+++ backend/modules/authentication/actions/index.php
@@ -60,6 +60,7 @@ class BackendAuthenticationIndex extends BackendBaseActionIndex
 	{
 		// assign the interface language ourself, because it won't be assigned automagically
 		$this->tpl->assign('INTERFACE_LANGUAGE', BackendLanguage::getInterfaceLanguage());
+
 		$this->frm->parse($this->tpl);
 		$this->frmForgotPassword->parse($this->tpl);
 	}